+++ jobis-application/src/main/java/team/retum/jobis/domain/application/usecase/QueryEmploymentRateUseCase.java
@@ -5,6 +5,7 @@ import team.retum.jobis.common.annotation.ReadOnlyUseCase;
 import team.retum.jobis.domain.company.dto.response.EmploymentRatesResponse;
 import team.retum.jobis.domain.company.spi.QueryCompanyPort;
 import team.retum.jobis.domain.company.spi.vo.CompanyVO;
+import team.retum.jobis.domain.student.spi.QueryStudentPort;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,6 +15,7 @@ import java.util.List;
 public class QueryEmploymentRateUseCase {
 
     private final QueryCompanyPort queryCompanyPort;
+    private final QueryStudentPort queryStudentPort;
     private static final Integer FIRTH_CLASS = 1;
     private static final Integer FOURTH_CLASS = 4;
 
@@ -22,7 +24,7 @@ public class QueryEmploymentRateUseCase {
 
         for (Integer classNum = FIRTH_CLASS; classNum <= FOURTH_CLASS; classNum++) {
             List<CompanyVO> companies = queryCompanyPort.getEmploymentRateByClassNumber(classNum);
-            Integer totalStudents = queryCompanyPort.getTotalStudentsByClassNumber(classNum);
+            Integer totalStudents = queryStudentPort.getTotalStudentsByClassNumber(classNum);
 
             classResponses.add(new EmploymentRatesResponse.ClassResponse(classNum, companies, totalStudents));
         }
