+++ jobis-application/src/main/java/team/retum/jobis/domain/company/spi/QueryCompanyPort.java
@@ -44,6 +44,4 @@ public interface QueryCompanyPort {
 
     List<CompanyVO> getEmploymentRateByClassNumber(Integer classNum);
 
-    Integer getTotalStudentsByClassNumber(Integer classNum);
-
 }
