+++ jobis-infrastructure/src/main/java/team/retum/jobis/domain/company/persistence/CompanyPersistenceAdapter.java
@@ -356,17 +356,6 @@ public class CompanyPersistenceAdapter implements CompanyPort {
             .fetch();
     }
 
-    @Override
-    public Integer getTotalStudentsByClassNumber(Integer classNum) {
-        Long count = queryFactory
-            .select(studentEntity.count())
-            .from(studentEntity)
-            .where(studentEntity.classRoom.eq(classNum))
-            .fetchOne();
-
-        return count != null ? count.intValue() : 0;
-    }
-
     private BooleanExpression containsName(String name) {
         return name == null ? null : companyEntity.name.contains(name);
     }
